extract tooltip placement
